Implements makefile and reviem print grid
