+++ Dockerfile
@@ -1,4 +1,4 @@
 FROM node
 ADD . /app
 WORKDIR /app
-RUN npm install
+RUN npm install && npm link
