+++ index.js
@@ -1,8 +1,10 @@
-const util = require('util')
+#!/usr/bin/env node
+
+const util = require("util");
 const readlineSync = require("readline-sync");
-const Commands = require('../lib/commands.js');
-const GridController = require('../controller/gridController.js');
-const Drone = require('../controller/droneController.js');
+const Commands = require('./src/lib/commands.js');
+const Grid = require('./src/lib/grid.js');
+const Drone = require('./src/lib/drone.js');
 
 command = new Commands;
 
@@ -17,7 +19,7 @@ if (!command.validateGrid(process.argv[2])){
 gridRange = command.partitionGrid(process.argv[2]);
 console.log(util.format('Generating flying grid with dimentions of %dm by %dm \n', gridRange.X, gridRange.Y));
 
-grid = new GridController([gridRange.X, gridRange.Y]);
+grid = new Grid([gridRange.X, gridRange.Y]);
 
 function report(droneId, finalPosition=[1,1], direction, photos) {
     console.log(util.format('- Drone %d:', droneId));                                                                                                                                                            
@@ -61,6 +63,8 @@ do {
 
                 });
                 drone = null;
+            } else if (droneUtils.X > gridRange.X || droneUtils.Y > gridRange.Y ) {
+                console.log(util.format('You cant start drone at [%d,%d], Grid out of range [%d,%d]', droneUtils.X, droneUtils.Y, gridRange.X, gridRange.Y))
             } else {
                 console.log(util.format('This drone can`t init in [%d, %d] because another drone init here.', droneUtils.X, droneUtils.Y));
                 console.log(util.format('For drone %d, try again!', droneCount));
@@ -82,5 +86,9 @@ while (repeat);
 
 // offer to user the option to print the grid
 if (readlineSync.keyInYN('Do you want to print GRID?')) {
-    console.log(grid.map);
+    grid.map.forEach(
+        function(line){
+            console.log(JSON.stringify(line, undefined, ''))
+
+    })
 }
