+++ run.sh
@@ -2,4 +2,4 @@
 #!/usr/bin/bash
 sudo docker build --no-cache -t mynode .
 clear
-sudo docker run -it mynode node src/view/interface.js $1
+sudo docker run -it mynode startGrid $1
