+++ src/lib/grid.js
@@ -1,6 +1,16 @@
-const Grid = require('../model/grid.js');
+class Grid {
+    constructor(mapRange = [1,1]) {
+        this._mapRange = [mapRange[0]-1, mapRange[1]-1]
+        this._map = [];
+        for (var y = 0; y <= mapRange[1]-1; y++) {
+            var columns = [];
+            for (var x = 0; x <= mapRange[0]-1; x++) {
+                columns[x] = '  ';
+            }
+            this._map[y] = columns;
+        }
+    }
 
-class GridController extends Grid {
     get map() {
         return this._map.reverse();
     }
@@ -20,4 +30,4 @@ class GridController extends Grid {
     }
 }
 
-module.exports = GridController;
+module.exports = Grid;
